test: TestPiece를 PieceTest의 내부 클래스로 이동
