+++ src/test/java/chess/domain/piece/PieceTest.java
@@ -1,5 +1,6 @@
 package chess.domain.piece;
 
+import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -23,5 +24,22 @@ class PieceTest {
         Piece piece2 = new TestPiece(Color.WHITE);
 
         assertThat(piece1.isSameColor(piece2)).isFalse();
+    }
+
+    private static class TestPiece extends Piece {
+
+        public TestPiece(Color color) {
+            super(color, ((fileDifference, rankDifference) -> false));
+        }
+
+        @Override
+        public boolean isMovable(Position from, Position to) {
+            return false;
+        }
+
+        @Override
+        public boolean isCatchable(Position from, Position to) {
+            return false;
+        }
     }
 }
